+++ service/controller/cloudconfig/master_template.go
@@ -200,19 +200,6 @@ func (me *masterExtension) Files() ([]k8scloudconfig.FileAsset, error) {
 			},
 			Permissions: FilePermission,
 		},
-		{
-			AssetContent: ignition.PythonPathProfile,
-			Path:         "/etc/profile.d/python.sh",
-			Owner: k8scloudconfig.Owner{
-				Group: k8scloudconfig.Group{
-					Name: FileOwnerGroupName,
-				},
-				User: k8scloudconfig.User{
-					Name: FileOwnerUserName,
-				},
-			},
-			Permissions: ReadAllFilePermission,
-		},
 	}
 
 	data := me.templateData(me.certFiles)
@@ -299,6 +286,11 @@ func (me *masterExtension) Units() ([]k8scloudconfig.UnitAsset, error) {
 		{
 			AssetContent: ignition.VNICConfigurationUnit,
 			Name:         "vnic-configuration.service",
+			Enabled:      true,
+		},
+		{
+			AssetContent: ignition.PythonExeLinkUnit,
+			Name:         "python-exe-link.service",
 			Enabled:      true,
 		},
 	}