+++ service/controller/cloudconfig/worker_template.go
@@ -89,19 +89,6 @@ func (we *workerExtension) Files() ([]k8scloudconfig.FileAsset, error) {
 			},
 			Permissions: CloudProviderFilePermission,
 		},
-		{
-			AssetContent: ignition.PythonPathProfile,
-			Path:         "/etc/profile.d/python.sh",
-			Owner: k8scloudconfig.Owner{
-				Group: k8scloudconfig.Group{
-					Name: FileOwnerGroupName,
-				},
-				User: k8scloudconfig.User{
-					Name: FileOwnerUserName,
-				},
-			},
-			Permissions: ReadAllFilePermission,
-		},
 	}
 
 	data := we.templateData(we.certFiles)
@@ -183,6 +170,11 @@ func (we *workerExtension) Units() ([]k8scloudconfig.UnitAsset, error) {
 		{
 			AssetContent: ignition.VNICConfigurationUnit,
 			Name:         "vnic-configuration.service",
+			Enabled:      true,
+		},
+		{
+			AssetContent: ignition.PythonExeLinkUnit,
+			Name:         "python-exe-link.service",
 			Enabled:      true,
 		},
 	}