+++ service/controller/templates/ignition/python_path.go
@@ -1,4 +1,10 @@
 package ignition
 
-const PythonPathProfile = `PATH=$PATH:/usr/share/oem/python/bin
+const PythonExeLinkUnit = `[Unit]
+Description=Link the Azure-provided python executable to make it available to MMA extension
+[Service]
+Type=oneshot
+ExecStart=/usr/bin/ln -s /usr/share/oem/python/bin/python /opt/bin/python2
+[Install]
+WantedBy=multi-user.target
 `